lab1 Refactor

+++ server/src/schemas/payment.py
@@ -1,5 +1,6 @@
 from pydantic import BaseModel, ConfigDict
 from datetime import datetime
+from decimal import Decimal
 from uuid import UUID
 
 from models.payment import PaymentStatus
@@ -9,7 +10,7 @@ class PaymentResponse(BaseModel):
     id: UUID
     lot_id: UUID
     user_id: UUID
-    amount: float
+    amount: Decimal
     status: PaymentStatus
     created_at: datetime
